fix(server): More fixes
